Added more hash tables, VTune integration
New hash tables: TBB (from github), libcds, folly, google sparsehash and densehash, some versions with our allocator

VTune integration means only insertion is measured

+++ dbs-ll.cpp
@@ -10,7 +10,7 @@ extern "C" {
 
 #include <hre/user.h>
 
-uint64_t MurmurHash64 (const void * key, int len, unsigned int seed) {
+uint64_t MurmurHash64 (const void * key, int len, uint64_t seed) {
     return MurmurHash64<>(key, len, seed);
 }
