+++ murmurhash.h
@@ -5,7 +5,7 @@
 #include <cstdint>
 
 template<typename K>
-uint64_t MurmurHash64 ( K * key, int len, unsigned int seed )
+uint64_t MurmurHash64 ( K * key, int len, uint64_t seed )
 {
     const uint64_t m = 0xc6a4a7935bd1e995;
     const int r = 47;
